Revert file-split
