--- massa-ledger-worker/src/ledger_db/helpers.rs
@@ -1,22 +1,0 @@
-/// For a given start prefix (inclusive), returns the correct end prefix (non-inclusive).
-/// This assumes the key bytes are ordered in lexicographical order.
-/// Since key length is not limited, for some case we return `None` because there is
-/// no bounded limit (every keys in the series `[]`, `[255]`, `[255, 255]` ...).
-pub(crate) fn end_prefix(prefix: &[u8]) -> Option<Vec<u8>> {
-    let mut end_range = prefix.to_vec();
-    while let Some(0xff) = end_range.last() {
-        end_range.pop();
-    }
-    if let Some(byte) = end_range.last_mut() {
-        *byte += 1;
-        Some(end_range)
-    } else {
-        None
-    }
-}
-
-#[test]
-fn test_end_prefix() {
-    assert_eq!(end_prefix(&[5, 6, 7]), Some(vec![5, 6, 8]));
-    assert_eq!(end_prefix(&[5, 6, 255]), Some(vec![5, 7]));
-}
